requirements.txt
annotated-types==0.7.0
anyio==4.6.0
asgiref==3.8.1
attrs==24.2.0
beautifulsoup4==4.12.3
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
chromedriver-autoinstaller==0.6.4
colorama==0.4.6
distro==1.9.0
Django==5.1.1
django-crispy-forms==2.3
django-environ==0.11.2
django-extensions==3.2.3
django-filter==24.3
django-jazzmin==3.0.1
djangorestframework==3.15.2
exceptiongroup==1.2.2
extensions==0.4
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
iniconfig==2.0.0
jiter==0.6.1
Markdown==3.7
markdown2==2.5.1
mysqlclient==2.2.5
openai==1.51.2
outcome==1.3.0.post0
packaging==24.1
pip-chill==1.0.3
pluggy==1.5.0
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
pyee==12.0.0
PyPDF2==3.0.1
PySocks==1.7.1
pytest==8.3.3
pytest-django==4.9.0
python-decouple==3.8
python-dotenv==1.0.1
requests==2.32.3
selenium==4.25.0
setuptools==75.1.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
sqlparse==0.5.1
tomli==2.0.2
tqdm==4.66.5
trio==0.26.2
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
webdriver-manager==4.0.2
websocket-client==1.8.0
wsproto==1.2.0
